// frontend/src/openapi/models/PlanEnum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overbae API
 * Backend API for the Overclaw / Overmind platform.
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * * `free` - free
 * * `pro` - pro
 * @export
 */
export const PlanEnum = {
    free: 'free',
    pro: 'pro'
} as const;
export type PlanEnum = typeof PlanEnum[keyof typeof PlanEnum];


export function instanceOfPlanEnum(value: any): boolean {
    for (const key in PlanEnum) {
        if (Object.prototype.hasOwnProperty.call(PlanEnum, key)) {
            if (PlanEnum[key as keyof typeof PlanEnum] === value) {
                return true;
            }
        }
    }
    return false;
}

export function PlanEnumFromJSON(json: any): PlanEnum {
    return PlanEnumFromJSONTyped(json, false);
}

export function PlanEnumFromJSONTyped(json: any, ignoreDiscriminator: boolean): PlanEnum {
    return json as PlanEnum;
}

export function PlanEnumToJSON(value?: PlanEnum | null): any {
    return value as any;
}

export function PlanEnumToJSONTyped(value: any, ignoreDiscriminator: boolean): PlanEnum {
    return value as PlanEnum;
}