// frontend/src/openapi/models/PlanUsage.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overbae API
 * Backend API for the Overclaw / Overmind platform.
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { PlanUsageUnit } from './PlanUsageUnit';
import {
    PlanUsageUnitFromJSON,
    PlanUsageUnitFromJSONTyped,
    PlanUsageUnitToJSON,
    PlanUsageUnitToJSONTyped,
} from './PlanUsageUnit';

/**
 *
 * @export
 * @interface PlanUsage
 */
export interface PlanUsage {
    /**
     *
     * @type {PlanUsageUnit}
     * @memberof PlanUsage
     */
    optimizeRuns: PlanUsageUnit;
    /**
     *
     * @type {PlanUsageUnit}
     * @memberof PlanUsage
     */
    trainingJobs: PlanUsageUnit;
    /**
     *
     * @type {PlanUsageUnit}
     * @memberof PlanUsage
     */
    deployJobs: PlanUsageUnit;
    /**
     *
     * @type {PlanUsageUnit}
     * @memberof PlanUsage
     */
    projects: PlanUsageUnit;
    /**
     *
     * @type {PlanUsageUnit}
     * @memberof PlanUsage
     */
    seats: PlanUsageUnit;
}

/**
 * Check if a given object implements the PlanUsage interface.
 */
export function instanceOfPlanUsage(value: object): value is PlanUsage {
    if (!('optimizeRuns' in value) || value['optimizeRuns'] === undefined) return false;
    if (!('trainingJobs' in value) || value['trainingJobs'] === undefined) return false;
    if (!('deployJobs' in value) || value['deployJobs'] === undefined) return false;
    if (!('projects' in value) || value['projects'] === undefined) return false;
    if (!('seats' in value) || value['seats'] === undefined) return false;
    return true;
}

export function PlanUsageFromJSON(json: any): PlanUsage {
    return PlanUsageFromJSONTyped(json, false);
}

export function PlanUsageFromJSONTyped(json: any, ignoreDiscriminator: boolean): PlanUsage {
    if (json == null) {
        return json;
    }
    return {

        'optimizeRuns': PlanUsageUnitFromJSON(json['optimize_runs']),
        'trainingJobs': PlanUsageUnitFromJSON(json['training_jobs']),
        'deployJobs': PlanUsageUnitFromJSON(json['deploy_jobs']),
        'projects': PlanUsageUnitFromJSON(json['projects']),
        'seats': PlanUsageUnitFromJSON(json['seats']),
    };
}

export function PlanUsageToJSON(json: any): PlanUsage {
    return PlanUsageToJSONTyped(json, false);
}

export function PlanUsageToJSONTyped(value?: PlanUsage | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'optimize_runs': PlanUsageUnitToJSON(value['optimizeRuns']),
        'training_jobs': PlanUsageUnitToJSON(value['trainingJobs']),
        'deploy_jobs': PlanUsageUnitToJSON(value['deployJobs']),
        'projects': PlanUsageUnitToJSON(value['projects']),
        'seats': PlanUsageUnitToJSON(value['seats']),
    };
}