// frontend/src/openapi/models/PlanUsageUnit.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overbae API
 * Backend API for the Overclaw / Overmind platform.
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface PlanUsageUnit
 */
export interface PlanUsageUnit {
    /**
     *
     * @type {number}
     * @memberof PlanUsageUnit
     */
    used?: number;
    /**
     *
     * @type {number}
     * @memberof PlanUsageUnit
     */
    limit: number | null;
}

/**
 * Check if a given object implements the PlanUsageUnit interface.
 */
export function instanceOfPlanUsageUnit(value: object): value is PlanUsageUnit {
    if (!('limit' in value) || value['limit'] === undefined) return false;
    return true;
}

export function PlanUsageUnitFromJSON(json: any): PlanUsageUnit {
    return PlanUsageUnitFromJSONTyped(json, false);
}

export function PlanUsageUnitFromJSONTyped(json: any, ignoreDiscriminator: boolean): PlanUsageUnit {
    if (json == null) {
        return json;
    }
    return {

        'used': json['used'] == null ? undefined : json['used'],
        'limit': json['limit'],
    };
}

export function PlanUsageUnitToJSON(json: any): PlanUsageUnit {
    return PlanUsageUnitToJSONTyped(json, false);
}

export function PlanUsageUnitToJSONTyped(value?: PlanUsageUnit | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'used': value['used'],
        'limit': value['limit'],
    };
}