// frontend/src/openapi/models/Subscription.ts
export interface Subscription {
    /**
     *
     * @type {boolean}
     * @memberof Subscription
     */
    billingEnabled: boolean;
    /**
     *
     * @type {PlanEnum}
     * @memberof Subscription
     */
    plan: PlanEnum;
    /**
     *
     * @type {string}
     * @memberof Subscription
     */
    status: string | null;
    /**
     *
     * @type {string}
     * @memberof Subscription
     */
    creditsUsd: string;
    /**
     *
     * @type {string}
     * @memberof Subscription
     */
    creditsGrantedUsd: string;
    /**
     *
     * @type {string}
     * @memberof Subscription
     */
    creditsSpentUsd: string;
    /**
     *
     * @type {Date}
     * @memberof Subscription
     */
    startDate: Date | null;
    /**
     *
     * @type {Date}
     * @memberof Subscription
     */
    endDate: Date | null;
    /**
     *
     * @type {boolean}
     * @memberof Subscription
     */
    cancelAtPeriodEnd: boolean;
    /**
     *
     * @type {PlanUsage}
     * @memberof Subscription
     */
    usage: PlanUsage;
}